requirements.txt
Django==3.2.2
requests==2.25.1
django-filter==2.4.0
djangorestframework==3.12.4
django-rest-auth==0.9.5
django-cors-headers==3.7.0
pytest-django==4.5.2
pytest-json-report==1.5.0